Split organization implementation files
