+++ desktop/src/components/settings/panes/OrganizationPane.tsx
@@ -5,7 +5,6 @@ import {
   useState,
   type ChangeEvent,
   type FormEvent,
-  type ReactNode,
 } from "react";
 import { useNavigate, useSearchParams } from "react-router-dom";
 import { Badge, type BadgeTone } from "@/components/ui/Badge";
@@ -16,6 +15,11 @@ import { CloudUpload, Mail, RefreshCw, Trash } from "@/components/ui/icons";
 import { SettingsCard } from "@/components/settings/SettingsCard";
 import { SettingsCardRow } from "@/components/settings/SettingsCardRow";
 import { SettingsPageHeader } from "@/components/settings/SettingsPageHeader";
+import {
+  Avatar,
+  OrganizationLogo,
+  OrganizationSection,
+} from "@/components/settings/panes/organization/OrganizationLogo";
 import { useActiveOrganization } from "@/hooks/organizations/use-active-organization";
 import { useOrganizationActions } from "@/hooks/organizations/use-organization-actions";
 import { useOrganizationInvitations } from "@/hooks/organizations/use-organization-invitations";
@@ -546,68 +550,5 @@ function InvitationRow({
         </div>
       ) : null}
     </div>
-  );
-}
-
-function OrganizationLogo({
-  organization,
-  logoImage,
-}: {
-  organization: OrganizationResponse;
-  logoImage?: string | null;
-}) {
-  const initials = organization.name.trim().slice(0, 2).toUpperCase() || "OR";
-  if (logoImage) {
-    return (
-      <div className="flex size-10 shrink-0 items-center justify-center overflow-hidden rounded-md bg-foreground/5">
-        <img src={logoImage} alt="" className="size-full object-cover" />
-      </div>
-    );
-  }
-  if (!organization.logoDomain) {
-    return (
-      <div className="flex size-10 shrink-0 items-center justify-center rounded-md bg-foreground/5 text-sm font-medium">
-        {initials}
-      </div>
-    );
-  }
-  const faviconUrl = `https://www.google.com/s2/favicons?domain=${encodeURIComponent(organization.logoDomain)}&sz=64`;
-  return (
-    <div className="flex size-10 shrink-0 items-center justify-center overflow-hidden rounded-md bg-foreground/5">
-      <img src={faviconUrl} alt="" className="size-6" />
-    </div>
-  );
-}
-
-function Avatar({ member }: { member: OrganizationMemberResponse }) {
-  const initials = (member.displayName || member.email).trim().slice(0, 2).toUpperCase();
-  if (member.avatarUrl) {
-    return (
-      <img
-        src={member.avatarUrl}
-        alt=""
-        className="size-8 rounded-full bg-foreground/5"
-      />
-    );
-  }
-  return (
-    <div className="flex size-8 shrink-0 items-center justify-center rounded-full bg-foreground/5 text-xs font-medium">
-      {initials || "U"}
-    </div>
-  );
-}
-
-function OrganizationSection({
-  title,
-  children,
-}: {
-  title: string;
-  children: ReactNode;
-}) {
-  return (
-    <section className="space-y-2">
-      <h2 className="text-sm font-medium text-foreground">{title}</h2>
-      {children}
-    </section>
   );
 }
